Handle resultCode "Received", add logging
When a payment request is not processed immediately, Adyen may return
the resultCode "Received". For these requests, we set the status to
depending on the "sdd_skip_closed" CiviSEPA setting. Processing of
payment updates (via Adyen's Notification API) will be done
separately.

This also adds logging of the exact API response.